Moved tasks new module. Added filter_tweets method
