--- inputs/insec_tweeters.csv
@@ -1,14 +1,0 @@
-MonieeLovee_
-itsjustdebb
-aqueenamongus
-JadoreJaggie
-britneybeesh
-AshleyKSmalls
-natasharothwell
-africamiranda
-SHEAUXBIZ07
-Complexchick
-TheRoot
-SylviaObell
-The_A_Prentice
-insecurehbo
